backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "fastapi-pagination>=0.15.6",
    "psycopg2-binary >=2.9.11",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
]

[tool.setuptools]
packages = ["app"]
